Move import playVoiceForTrack into players/actions.js. Dispatch action to invoke playback from middleware
