+++ app/player/middleware.js
@@ -4,10 +4,13 @@ export default (store) => (next) => (action) => {
   switch (action.type) {
     case 'REGISTER_PLAYER':
       playersByTrackId[action.trackId] = action.player
-      break;
+      return
     case 'UNREGISTER_PLAYER':
       delete playersByTrackId[action.trackId]
-      break;
+      return
+    case 'PLAYER_PLAY':
+      playersByTrackId[action.trackId] && playersByTrackId[action.trackId].play(action.playbackData)
+      return
   }
 
   return next(action)
