+++ app/voices/ChromaticStepEditorContainer.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { updateSelectedStepPitch } from '../sequencer/actions'
-import { playVoiceForTrack } from './actions'
+import { playVoiceForTrack } from '../player/actions'
 import ChromaticKeyboardPad from './ChromaticKeyboardPad'
 
 const ChromaticStepEditorContainer = (props) => <ChromaticKeyboardPad {...props} />
